requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
sqlalchemy==2.0.27
aiosqlite==0.19.0
pillow==10.2.0
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1 